One EC client instead of two
