+++ src/main/scala/units/ConsensusClient.scala
@@ -16,7 +16,6 @@ import net.ceedubs.ficus.Ficus.*
 import org.slf4j.LoggerFactory
 import sttp.client3.HttpClientSyncBackend
 import units.client.engine.{EngineApiClient, HttpEngineApiClient}
-import units.client.http.{EcApiClient, HttpEcApiClient}
 import units.client.{JwtAuthenticationBackend, LoggingBackend}
 import units.network.*
 
@@ -28,7 +27,6 @@ class ConsensusClient(
     config: ClientConfig,
     context: ExtensionContext,
     engineApiClient: EngineApiClient,
-    httpApiClient: EcApiClient,
     blockObserver: BlocksObserver,
     allChannels: DefaultChannelGroup,
     globalScheduler: Scheduler,
@@ -42,7 +40,6 @@ class ConsensusClient(
       deps.config,
       context,
       deps.engineApiClient,
-      deps.httpApiClient,
       deps.blockObserver,
       deps.allChannels,
       deps.globalScheduler,
@@ -54,7 +51,6 @@ class ConsensusClient(
 
   private[units] val elu =
     new ELUpdater(
-      httpApiClient,
       engineApiClient,
       context.blockchain,
       context.utx,
@@ -123,14 +119,13 @@ class ConsensusClientDependencies(context: ExtensionContext) extends AutoCloseab
         httpClientBackend
     }
   )
-  val httpApiClient = new HttpEcApiClient(config, httpClientBackend)
 
   val allChannels     = new DefaultChannelGroup(GlobalEventExecutor.INSTANCE)
   val peerDatabase    = new PeerDatabaseImpl(config.network)
   val messageObserver = new MessageObserver()
   private val networkServer = NetworkServer(
     config,
-    new HistoryReplier(httpApiClient, engineApiClient)(globalScheduler),
+    new HistoryReplier(engineApiClient)(globalScheduler),
     peerDatabase,
     messageObserver,
     allChannels,
