+++ src/main/scala/units/client/engine/EngineApiClient.scala
@@ -1,13 +1,12 @@
 package units.client.engine
 
+import play.api.libs.json.*
 import units.client.engine.EngineApiClient.PayloadId
 import units.client.engine.model.*
 import units.eth.EthAddress
 import units.{BlockHash, Job}
-import play.api.libs.json.*
 
 trait EngineApiClient {
-
   def forkChoiceUpdate(blockHash: BlockHash, finalizedBlockHash: BlockHash): Job[String] // TODO Replace String with an appropriate type
 
   def forkChoiceUpdateWithPayloadId(
@@ -24,6 +23,18 @@ trait EngineApiClient {
   def applyNewPayload(payload: JsObject): Job[Option[BlockHash]]
 
   def getPayloadBodyByHash(hash: BlockHash): Job[Option[JsObject]]
+
+  def getBlockByNumber(number: BlockNumber): Job[Option[EcBlock]]
+
+  def getBlockByHash(hash: BlockHash): Job[Option[EcBlock]]
+
+  def getBlockByHashJson(hash: BlockHash, fullTxs: Boolean = false): Job[Option[JsObject]]
+
+  def getLastExecutionBlock: Job[EcBlock]
+
+  def blockExists(hash: BlockHash): Job[Boolean]
+
+  def getLogs(hash: BlockHash, topic: String): Job[List[GetLogsResponseEntry]]
 }
 
 object EngineApiClient {
