+++ src/main/scala/units/client/engine/model/GetBlockByHashRequest.scala
@@ -1,4 +1,4 @@
-package units.client.http.model
+package units.client.engine.model
 
 import units.BlockHash
 import play.api.libs.json.{Json, Writes}
