+++ src/main/scala/units/client/engine/model/GetBlockByNumberRequest.scala
@@ -1,4 +1,4 @@
-package units.client.http.model
+package units.client.engine.model
 
 import play.api.libs.json.{Json, Writes}
