+++ src/main/scala/units/client/engine/model/GetLogsResponseEntry.scala
@@ -1,4 +1,4 @@
-package units.client.http.model
+package units.client.engine.model
 
 import units.eth.EthAddress
 import play.api.libs.json.{Json, Reads}
