+++ src/main/scala/units/network/HistoryReplier.scala
@@ -1,24 +1,20 @@
 package units.network
 
 import cats.syntax.either.*
-import units.client.engine.EngineApiClient
-import units.client.http.EcApiClient
-import units.util.BlockToPayloadMapper
-import units.{BlockHash, ClientError, NetworkL2Block}
 import com.wavesplatform.network.id
 import com.wavesplatform.utils.ScorexLogging
 import io.netty.channel.ChannelHandler.Sharable
 import io.netty.channel.{ChannelHandlerContext, ChannelInboundHandlerAdapter}
 import monix.execution.Scheduler
+import units.client.engine.EngineApiClient
+import units.util.BlockToPayloadMapper
+import units.{BlockHash, ClientError, NetworkL2Block}
 
 import scala.concurrent.Future
 import scala.util.{Failure, Success}
 
 @Sharable
-class HistoryReplier(httpApiClient: EcApiClient, engineApiClient: EngineApiClient)(implicit sc: Scheduler)
-    extends ChannelInboundHandlerAdapter
-    with ScorexLogging {
-
+class HistoryReplier(engineApiClient: EngineApiClient)(implicit sc: Scheduler) extends ChannelInboundHandlerAdapter with ScorexLogging {
   private def respondWith(ctx: ChannelHandlerContext, value: Future[Message]): Unit =
     value.onComplete {
       case Failure(e) => log.debug(s"${id(ctx)} Error processing request", e)
@@ -46,7 +42,7 @@ class HistoryReplier(httpApiClient: EcApiClient, engineApiClient: EngineApiClien
 
   private def loadBlockL2(hash: BlockHash): Future[Either[ClientError, NetworkL2Block]] = Future {
     for {
-      blockJsonOpt       <- httpApiClient.getBlockByHashJson(hash)
+      blockJsonOpt       <- engineApiClient.getBlockByHashJson(hash)
       blockJson          <- Either.fromOption(blockJsonOpt, ClientError("block not found"))
       payloadBodyJsonOpt <- engineApiClient.getPayloadBodyByHash(hash)
       payloadBodyJson    <- Either.fromOption(payloadBodyJsonOpt, ClientError("payload body not found"))
