+++ src/test/scala/units/BaseIntegrationTestSuite.scala
@@ -11,7 +11,7 @@ import com.wavesplatform.utils.ScorexLogging
 import org.scalatest.freespec.AnyFreeSpec
 import org.scalatest.{BeforeAndAfterAll, EitherValues, OptionValues}
 import units.Bridge.ElSentNativeEvent
-import units.client.http.model.GetLogsResponseEntry
+import units.client.engine.model.GetLogsResponseEntry
 import units.eth.{EthAddress, Gwei}
 import units.test.CustomMatchers
 import units.util.HexBytesConverter
