+++ src/test/scala/units/client/engine/LoggedEngineApiClient.scala
@@ -2,7 +2,7 @@ package units.client.engine
 
 import play.api.libs.json.JsObject
 import units.client.engine.EngineApiClient.PayloadId
-import units.client.engine.model.Withdrawal
+import units.client.engine.model.*
 import units.eth.EthAddress
 import units.{BlockHash, HasJobLogging, Job}
 
@@ -29,6 +29,20 @@ class LoggedEngineApiClient(underlying: EngineApiClient) extends EngineApiClient
 
   override def getPayloadBodyByHash(hash: BlockHash): Job[Option[JsObject]] =
     wrap(s"getPayloadBodyByHash($hash)", underlying.getPayloadBodyByHash(hash))
+
+  override def getBlockByNumber(number: BlockNumber): Job[Option[EcBlock]] = wrap(s"getBlockByNumber($number)", underlying.getBlockByNumber(number))
+
+  override def getBlockByHash(hash: BlockHash): Job[Option[EcBlock]] = wrap(s"getBlockByHash($hash)", underlying.getBlockByHash(hash))
+
+  override def getBlockByHashJson(hash: BlockHash, fullTxs: Boolean): Job[Option[JsObject]] =
+    wrap(s"getBlockByHashJson($hash, fullTxs=$fullTxs)", underlying.getBlockByHashJson(hash, fullTxs))
+
+  override def getLastExecutionBlock: Job[EcBlock] = wrap("getLastExecutionBlock", underlying.getLastExecutionBlock)
+
+  override def blockExists(hash: BlockHash): Job[Boolean] = wrap(s"blockExists($hash)", underlying.blockExists(hash))
+
+  override def getLogs(hash: BlockHash, topic: String): Job[List[GetLogsResponseEntry]] =
+    wrap(s"getLogs($hash, $topic)", underlying.getLogs(hash, topic))
 }
 
 object LoggedEngineApiClient {
