+++ src/test/scala/units/client/engine/model/TestEcBlocks.scala
@@ -1,4 +1,4 @@
-package units.client.http.model
+package units.client.engine.model
 
 import com.wavesplatform.account.SeedKeyPair
 import com.wavesplatform.common.utils.EitherExt2
